some click_cli moved to utils. Fixed click_cli tests
