+++ nameko/click_cli/shell.py
@@ -2,49 +2,52 @@ import os
 import sys
 from types import ModuleType
 
-import nameko.click_cli.code as code
 from nameko import config
 from nameko.constants import AMQP_URI_CONFIG_KEY
 from nameko.standalone.events import event_dispatcher
 from nameko.standalone.rpc import ClusterRpcClient
 
-from .commands import Shell
+from .utils import code
+
+
+SHELLS = ["bpython", "ipython", "plain"]
 
 
 class ShellRunner(object):
-
     def __init__(self, banner, local):
         self.banner = banner
         self.local = local
 
     def bpython(self):
         import bpython  # pylint: disable=E0401
+
         bpython.embed(banner=self.banner, locals_=self.local)
 
     def ipython(self):
         from IPython import embed  # pylint: disable=E0401
+
         embed(banner1=self.banner, user_ns=self.local)
 
     def plain(self):
         code.interact(
             banner=self.banner,
             local=self.local,
-            raise_expections=not sys.stdin.isatty()
+            raise_expections=not sys.stdin.isatty(),
         )
 
     def start_shell(self, name):
         if not sys.stdin.isatty():
-            available_shells = ['plain']
+            available_shells = ["plain"]
         else:
-            available_shells = [name] if name else Shell.SHELLS
+            available_shells = [name] if name else SHELLS
 
         # Support the regular Python interpreter startup script if someone
         # is using it.
-        startup = os.environ.get('PYTHONSTARTUP')
+        startup = os.environ.get("PYTHONSTARTUP")
         if startup and os.path.isfile(startup):
-            with open(startup, 'r') as f:
-                eval(compile(f.read(), startup, 'exec'), self.local)
-            del os.environ['PYTHONSTARTUP']
+            with open(startup, "r") as f:
+                eval(compile(f.read(), startup, "exec"), self.local)
+            del os.environ["PYTHONSTARTUP"]
 
         for name in available_shells:
             try:
@@ -58,7 +61,7 @@ def make_nameko_helper():
     """Create a fake module that provides some convenient access to nameko
     standalone functionality for interactive shell usage.
     """
-    module = ModuleType('nameko')
+    module = ModuleType("nameko")
     module.__doc__ = """Nameko shell helper for making rpc calls and dispatching
 events.
 
@@ -78,14 +81,14 @@ Usage:
 
 def main(interface):
 
-    banner = 'Nameko Python %s shell on %s\nBroker: %s' % (
+    banner = "Nameko Python %s shell on %s\nBroker: %s" % (
         sys.version,
         sys.platform,
         config[AMQP_URI_CONFIG_KEY],
     )
 
     ctx = {}
-    ctx['n'] = make_nameko_helper()
+    ctx["n"] = make_nameko_helper()
 
     runner = ShellRunner(banner, ctx)
     # import pdb; pdb.set_trace()
