+++ test/click_cli/conftest.py
@@ -6,7 +6,7 @@ from mock import patch
 
 @pytest.fixture
 def command():
-    from nameko.click_cli.main import main
+    from nameko.click_cli import main
 
     def _command(*argv):
         with patch.object(sys, "argv", list(argv)):
