+++ test/click_cli/test_crun.py
@@ -10,9 +10,9 @@ import pytest
 from mock import patch
 
 from nameko import config
-from nameko.cli.run import import_service, run, setup_backdoor
+from nameko.click_cli.utils.import_services import import_services
+from nameko.click_cli.run import run, setup_backdoor
 from nameko.constants import SERIALIZER_CONFIG_KEY, WEB_SERVER_CONFIG_KEY
-from nameko.exceptions import CommandError
 from nameko.runners import ServiceRunner
 from nameko.standalone.rpc import ClusterRpcClient
 from nameko.testing.waiting import wait_for_call
@@ -56,7 +56,7 @@ def test_main_with_config(command, tmpdir):
     assert WEB_SERVER_CONFIG_KEY not in config
     assert SERIALIZER_CONFIG_KEY not in config
 
-    with patch("nameko.click_cli.run.run") as run:
+    with patch("nameko.click_cli.main_run") as run:
 
         command("cnameko", "run", "--config", config_file.strpath, "test.sample")
 
@@ -81,7 +81,7 @@ def test_main_with_config_options(command, tmpdir):
     assert SERIALIZER_CONFIG_KEY not in config
     assert "EGG" not in config
 
-    with patch("nameko.click_cli.run.run") as run:
+    with patch("nameko.click_cli.main_run") as run:
 
         command(
             "cnameko",
@@ -152,43 +152,43 @@ def test_main_with_logging_config(command, tmpdir):
 
 
 def test_import_ok():
-    assert import_service("test.sample") == [Service]
-    assert import_service("test.sample:Service") == [Service]
+    assert import_services("test.sample") == [Service]
+    assert import_services("test.sample:Service") == [Service]
 
 
 def test_import_missing():
-    with pytest.raises(CommandError) as exc:
-        import_service("non_existent")
+    with pytest.raises(ValueError) as exc:
+        import_services("non_existent")
     assert "No module named" in str(exc.value)
     assert "non_existent" in str(exc.value)
 
 
 def test_import_filename():
-    with pytest.raises(CommandError) as exc:
-        import_service("test/sample.py")
+    with pytest.raises(ValueError) as exc:
+        import_services("test/sample.py")
     assert "did you mean 'test.sample'?" in str(exc)
 
 
 def test_import_broken():
     with pytest.raises(ImportError):
-        import_service("test.broken_sample")
+        import_services("test.broken_sample")
 
 
 def test_import_missing_class():
-    with pytest.raises(CommandError) as exc:
-        import_service("test.sample:NonExistent")
+    with pytest.raises(ValueError) as exc:
+        import_services("test.sample:NonExistent")
     assert "Failed to find service class" in str(exc)
 
 
 def test_import_not_a_class():
-    with pytest.raises(CommandError) as exc:
-        import_service("test.sample:rpc")
+    with pytest.raises(ValueError) as exc:
+        import_services("test.sample:rpc")
     assert "Service must be a class" in str(exc)
 
 
 def test_import_no_service_classes():
-    with pytest.raises(CommandError):
-        import_service("test")
+    with pytest.raises(ValueError):
+        import_services("test")
 
 
 def recv_until_prompt(sock):
@@ -202,7 +202,7 @@ def recv_until_prompt(sock):
 
 def test_backdoor():
     runner = object()
-    green_socket, gt = setup_backdoor(runner, 0)
+    green_socket, gt = setup_backdoor(runner, ("localhost", 0))
     eventlet.sleep(0)  # give backdoor a chance to spawn
     socket_name = green_socket.fd.getsockname()
     sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
@@ -221,7 +221,7 @@ def test_backdoor():
 
 
 def test_stopping(rabbit_config):
-    with patch("nameko.cli.run.eventlet") as mock_eventlet:
+    with patch("nameko.click_cli.run.eventlet") as mock_eventlet:
         # this is the service "runlet"
         mock_eventlet.spawn().wait.side_effect = [
             KeyboardInterrupt,
@@ -233,13 +233,13 @@ def test_stopping(rabbit_config):
 
 
 def test_stopping_twice(rabbit_config):
-    with patch("nameko.cli.run.eventlet") as mock_eventlet:
+    with patch("nameko.click_cli.run.eventlet") as mock_eventlet:
         # this is the service "runlet"
         mock_eventlet.spawn().wait.side_effect = [
             KeyboardInterrupt,
             None,  # second wait, after stop() which returns normally
         ]
-        with patch("nameko.cli.run.ServiceRunner") as runner_cls:
+        with patch("nameko.click_cli.run.ServiceRunner") as runner_cls:
             runner = runner_cls()
             runner.stop.side_effect = KeyboardInterrupt
             runner.kill.return_value = None
@@ -249,7 +249,7 @@ def test_stopping_twice(rabbit_config):
 
 
 def test_os_error_for_signal(rabbit_config):
-    with patch("nameko.cli.run.eventlet") as mock_eventlet:
+    with patch("nameko.click_cli.run.eventlet") as mock_eventlet:
         # this is the service "runlet"
         mock_eventlet.spawn().wait.side_effect = [
             OSError(errno.EINTR, ""),
@@ -264,7 +264,7 @@ def test_os_error_for_signal(rabbit_config):
 
 
 def test_other_errors_propagate(rabbit_config):
-    with patch("nameko.cli.run.eventlet") as mock_eventlet:
+    with patch("nameko.click_cli.run.eventlet") as mock_eventlet:
         # this is the service "runlet"
         mock_eventlet.spawn().wait.side_effect = [
             OSError(0, ""),
@@ -282,7 +282,7 @@ def test_other_errors_propagate(rabbit_config):
 def test_broker_option_deprecated(command, rabbit_uri):
 
     with patch("nameko.click_cli.run.run") as run:
-        with patch("nameko.click_cli.config.warnings") as warnings:
+        with patch("nameko.click_cli.utils.config.warnings") as warnings:
             command("cnameko", "run", "--broker", rabbit_uri, "test.sample")
 
     assert run.call_count == 1
